+++ ldap2pg/role.py
@@ -142,15 +142,7 @@ class Role(object):
                 )
             )
 
-    _drop_objects_sql = dedent("""
-    DO $$BEGIN EXECUTE 'GRANT "%(role)s" TO "'||SESSION_USER||'";'; END$$;
-    DO $$BEGIN
-      EXECUTE 'REASSIGN OWNED BY "%(role)s" TO "'||SESSION_USER||'";';
-    END$$;
-    DROP OWNED BY "%(role)s";
-    """)
-
-    def drop(self):
+    def drop(self, databases=None):
         yield Query(
             'Terminate running session for %s.' % self.name,
             None, dedent("""\
@@ -159,11 +151,22 @@ class Role(object):
             WHERE usename = '%s';
             """) % self.name,
         )
-        yield Query(
-            'Reassign %s objects and purge ACL on %%(dbname)s.' % (self.name,),
-            Query.ALL_DATABASES,
-            self._drop_objects_sql % dict(role=self.name),
-        )
+        databases = databases or []
+        for db in databases:
+            yield Query(
+                "Reassign %s's objects in %s." % (self.name, db),
+                db.name,
+                dedent("""\
+                REVOKE "%(owner)s" FROM "%(role)s";
+                GRANT "%(role)s" TO "%(owner)s";
+                REASSIGN OWNED BY "%(role)s" TO "%(owner)s";
+                """) % dict(role=self.name, owner=db.owner),
+            )
+            yield Query(
+                "Purge %s's ACL in %s." % (self.name, db),
+                db.name,
+                """DROP OWNED BY "%(role)s";""" % dict(role=self.name),
+            )
         yield Query(
             'Drop %s.' % (self.name,),
             None,
@@ -326,7 +329,9 @@ class RoleSet(set):
     def union(self, other):
         return self.__class__(self | other)
 
-    def diff(self, other=None, available=None):
+    def diff(
+            self, other=None, available=None, fallback_owner=None,
+            databases=None):
         # Yield query so that self match other. It's kind of a three-way diff
         # since we reuse `available` roles instead of recreating roles.
 
@@ -434,8 +439,16 @@ class RoleSet(set):
 
         # Don't forget to trash all spurious managed roles!
         spurious = RoleSet(self - other - set(['public']))
+        # reassign databases to fallback_owner
+        for database in databases or []:
+            if database.owner in spurious:
+                for query in database.reassign(fallback_owner):
+                    yield query
+                # Update representation for the following queries.
+                database.owner = fallback_owner
+
         for role in reversed(list(spurious.flatten())):
-            for qry in role.drop():
+            for qry in role.drop(databases or []):
                 yield qry